painter cleanup
