+++ app/packages/looker/src/numpy.ts
@@ -16,6 +16,8 @@ export const ARRAY_TYPES = {
   Int32Array,
   Float32Array,
   Float64Array,
+  BigUint64Array,
+  BigInt64Array,
 };
 
 export type TypedArray =
@@ -26,7 +28,9 @@ export type TypedArray =
   | Uint32Array
   | Int32Array
   | Float32Array
-  | Float64Array;
+  | Float64Array
+  | BigUint64Array
+  | BigInt64Array;
 
 export interface OverlayMask {
   buffer: ArrayBuffer;
@@ -59,10 +63,10 @@ const DATA_TYPES = {
   "|i4": Int32Array,
   "<i4": Int32Array,
 
-  "|u8": convert64to32Array(Uint32Array),
-  "<u8": convert64to32Array(Uint32Array),
-  "|i8": convert64to32Array(Int32Array),
-  "<i8": convert64to32Array(Int32Array),
+  "|u8": BigUint64Array,
+  "<u8": BigUint64Array,
+  "|i8": BigInt64Array,
+  "<i8": BigInt64Array,
 
   "<f4": Float32Array,
   "|f4": Float32Array,
@@ -70,35 +74,6 @@ const DATA_TYPES = {
   "<f8": Float64Array,
   "|f8": Float64Array,
 };
-
-/**
- * Polyfill to convert a 64-bit integer array to a 32-bit integer array. This
- * assumes that no element actually requires more than 32 bits to store, which
- * should be a safe assumption for our purposes.
- */
-function convert64to32Array(
-  TargetArrayType: typeof Uint32Array | typeof Int32Array
-) {
-  // we only need the 3-argument constructor to be implemented. For details:
-  // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/TypedArray
-  const makeArray = function (
-    buffer: ArrayBuffer,
-    byteOffset: number,
-    length: number
-  ) {
-    // view buffer as 32-bit type and copy the 4 lowest bytes out of every 8
-    // bytes into a new array (assumes little-endian)
-    const source = new TargetArrayType(buffer, byteOffset, length * 2);
-    const target = new TargetArrayType(source.length);
-    for (let i = 0; i < target.length; i++) {
-      target[i] = source[i * 2];
-    }
-    return target;
-  };
-  // needed by parse()
-  makeArray.BYTES_PER_ELEMENT = 8;
-  return makeArray;
-}
 
 /**
  * Parses a uint16 (unsigned 16-bit integer) at a specified position in a
@@ -158,6 +133,12 @@ function parse(array: Uint8Array): OverlayMask {
           rawData.byteLength / ArrayType.BYTES_PER_ELEMENT
         );
 
+  console.log({
+    arrayType: typedData.constructor.name,
+    buffer: typedData.buffer,
+    channels: header.shape[2] ?? 1,
+    shape: [header.shape[0], header.shape[1]],
+  });
   return {
     arrayType: typedData.constructor.name,
     buffer: typedData.buffer,
