+++ app/packages/looker/src/overlays/segmentation.ts
@@ -3,19 +3,14 @@
  */
 
 import { getColor } from "@fiftyone/utilities";
-import { ARRAY_TYPES, OverlayMask, TypedArray } from "../numpy";
-import { BaseState, Coordinates, MaskTargets } from "../state";
-import {
-  BaseLabel,
-  CONTAINS,
-  Overlay,
-  PointInfo,
-  SelectData,
-  isShown,
-} from "./base";
+import type { OverlayMask, TypedArray } from "../numpy";
+import { ARRAY_TYPES } from "../numpy";
+import type { BaseState, Coordinates, MaskTargets } from "../state";
+import type { BaseLabel, Overlay, PointInfo, SelectData } from "./base";
+import { CONTAINS, isShown } from "./base";
 import { isRgbMaskTargets, sizeBytes, strokeCanvasRect, t } from "./util";
 
-interface SegmentationLabel extends BaseLabel {
+export interface SegmentationLabel extends BaseLabel {
   mask?: {
     data: OverlayMask;
     image: ArrayBuffer;
@@ -125,26 +120,7 @@ export default class SegmentationOverlay<State extends BaseState>
     if (this.containsPoint(state)) {
       return 0;
     }
-    return Infinity;
-  }
-
-  initRgbMaskTargetsCache(rgbMaskTargets: MaskTargets) {
-    if (!this.isRgbMaskTargets && isRgbMaskTargets(rgbMaskTargets)) {
-      this.isRgbMaskTargets = true;
-    }
-
-    if (this.rgbMaskTargetsReverseMap) {
-      return;
-    }
-
-    this.rgbMaskTargetsReverseMap = {};
-
-    Object.entries(rgbMaskTargets).map(([color, intTargetAndLabel]) => {
-      this.rgbMaskTargetsReverseMap[intTargetAndLabel.intTarget] = {
-        color,
-        label: intTargetAndLabel.label,
-      };
-    });
+    return Number.Infinity;
   }
 
   getPointInfo(state: Readonly<State>): Partial<PointInfo<SegmentationInfo>> {
@@ -161,76 +137,35 @@ export default class SegmentationOverlay<State extends BaseState>
     const target = this.getTarget(state);
 
     if (this.label.mask.data.channels > 2) {
-      const rgbSegmentationInfoWithoutColor: Omit<
-        PointInfo<SegmentationInfo>,
-        "color"
-      > = {
-        label: {
-          ...this.label,
-          mask: {
-            shape: this.label.mask.data.shape
-              ? this.label.mask.data.shape
-              : null,
-          },
-        },
-        field: this.field,
-        target,
-        type: "Segmentation",
-      };
-
-      this.initRgbMaskTargetsCache(maskTargets);
-
-      if (!this.isRgbMaskTargets) {
-        // getting color here is computationally inefficient, return no tooltip ribbon color for this edge case
-        return rgbSegmentationInfoWithoutColor;
-      }
-
-      if (
-        !this.rgbMaskTargetsReverseMap ||
-        !this.rgbMaskTargetsReverseMap[target]
-      ) {
-        return undefined;
-      }
-
-      return {
-        color: this.rgbMaskTargetsReverseMap[target].color,
-        label: {
-          ...this.label,
-          mask: {
-            shape: this.label.mask.data.shape
-              ? this.label.mask.data.shape
-              : null,
-          },
-        },
-        field: this.field,
-        target,
-        type: "Segmentation",
-      };
-    } else {
-      return {
-        color:
-          maskTargets && Object.keys(maskTargets).length === 1
-            ? getColor(
-                state.options.coloring.pool,
-                state.options.coloring.seed,
-                this.field
-              )
-            : coloring.targets[
-                Math.round(Math.abs(target)) % coloring.targets.length
-              ],
-        label: {
-          ...this.label,
-          mask: {
-            shape: this.label.mask.data.shape
-              ? this.label.mask.data.shape
-              : null,
-          },
-        },
-        field: this.field,
-        target,
-        type: "Segmentation",
-      };
+      return this.handleRgbTargets(maskTargets, target);
     }
+
+    if (this.label.mask.data.channels === 2) {
+      console.log(target);
+      throw new Error("E");
+    }
+
+    return {
+      color:
+        maskTargets && Object.keys(maskTargets).length === 1
+          ? getColor(
+              state.options.coloring.pool,
+              state.options.coloring.seed,
+              this.field
+            )
+          : coloring.targets[
+              Math.round(Math.abs(target)) % coloring.targets.length
+            ],
+      label: {
+        ...this.label,
+        mask: {
+          shape: this.label.mask.data.shape ? this.label.mask.data.shape : null,
+        },
+      },
+      field: this.field,
+      target,
+      type: "Segmentation",
+    };
   }
 
   getSelectData(): SelectData {
@@ -280,12 +215,76 @@ export default class SegmentationOverlay<State extends BaseState>
     if (index < 0) {
       return null;
     }
-    return this.targets[index];
+    return Number(this.targets[index]);
+  }
+
+  private handleRgbTargets(targets: MaskTargets, target: number) {
+    const rgbSegmentationInfoWithoutColor: Omit<
+      PointInfo<SegmentationInfo>,
+      "color"
+    > = {
+      label: {
+        ...this.label,
+        mask: {
+          shape: this.label.mask.data.shape ? this.label.mask.data.shape : null,
+        },
+      },
+      field: this.field,
+      target,
+      type: "Segmentation",
+    };
+
+    this.initRgbMaskTargetsCache(targets);
+
+    if (!this.isRgbMaskTargets) {
+      // getting color here is computationally inefficient, return no tooltip
+      // ribbon color for this edge case
+      return rgbSegmentationInfoWithoutColor;
+    }
+
+    if (
+      !this.rgbMaskTargetsReverseMap ||
+      !this.rgbMaskTargetsReverseMap[target]
+    ) {
+      return undefined;
+    }
+
+    return {
+      color: this.rgbMaskTargetsReverseMap[target].color,
+      label: {
+        ...this.label,
+        mask: {
+          shape: this.label.mask.data.shape ? this.label.mask.data.shape : null,
+        },
+      },
+      field: this.field,
+      target,
+      type: "Segmentation",
+    };
+  }
+
+  private initRgbMaskTargetsCache(targets: MaskTargets) {
+    if (!this.isRgbMaskTargets && isRgbMaskTargets(targets)) {
+      this.isRgbMaskTargets = true;
+    }
+
+    if (this.rgbMaskTargetsReverseMap) {
+      return;
+    }
+
+    this.rgbMaskTargetsReverseMap = {};
+
+    Object.entries(targets).map(([color, intTargetAndLabel]) => {
+      this.rgbMaskTargetsReverseMap[intTargetAndLabel.intTarget] = {
+        color,
+        label: intTargetAndLabel.label,
+      };
+    });
   }
 }
 
 export const getSegmentationPoints = (
-  labels: SegmentationLabel[]
+  _: SegmentationLabel[]
 ): Coordinates[] => {
   return [
     [0, 0],
