+++ app/packages/looker/src/worker/indexed-png-decoder.test.ts
@@ -1,9 +1,7 @@
+import type { BitDepth } from "fast-png";
 import { describe, expect, it } from "vitest";
-import {
-  ColorPalette,
-  ValidPngBitDepth,
-  indexedPngBufferToRgb,
-} from "../indexed-png-decoder";
+import type { ColorPalette } from "./indexed-png-decoder";
+import { indexedPngBufferToRgb } from "./indexed-png-decoder";
 
 describe("convertIndexedPngToRgb", () => {
   const palette: ColorPalette = [
@@ -15,7 +13,7 @@ describe("convertIndexedPngToRgb", () => {
 
   it("works with 1-bit depth", () => {
     const inputData = new Uint8Array([0b10000000]);
-    const bitDepth: ValidPngBitDepth = 1;
+    const bitDepth: BitDepth = 1;
     // only two colors possible with 1 bit depth
     // for 0b10000000, only first bit has value, so the first index color (1, 1, 1) is chosen
     // other bits get 0 indexed color (0, 0, 0)
@@ -29,7 +27,7 @@ describe("convertIndexedPngToRgb", () => {
 
   it("works with 2-bit depth", () => {
     const inputData = new Uint8Array([0b11100000]);
-    const bitDepth: ValidPngBitDepth = 2;
+    const bitDepth: BitDepth = 2;
     // four colors possible with 2 bit depth
     // first two bits = 11 = 3, so the third index color (3, 3, 3) is chosen
     // second two bits = 10 = 2, so the second color is chosen
@@ -42,7 +40,7 @@ describe("convertIndexedPngToRgb", () => {
 
   it("works with 4-bit depth", () => {
     const inputData = new Uint8Array([0b00110010]);
-    const bitDepth: ValidPngBitDepth = 4;
+    const bitDepth: BitDepth = 4;
     // 0011 = 3, so the third index color (3, 3, 3) is chosen
     // 0010 = 2, so the second index color is chosen
     const expectedResult = new Uint8Array([3, 3, 3, 2, 2, 2]);
@@ -53,7 +51,7 @@ describe("convertIndexedPngToRgb", () => {
 
   it("works with 8-bit depth", () => {
     const inputData = new Uint8Array([0b00000010, 0b00000011]);
-    const bitDepth: ValidPngBitDepth = 8;
+    const bitDepth: BitDepth = 8;
     // 00000010 = 2, so the second index color (2, 2, 2) is chosen
     // 00000011 = 3, so the third index color (3, 3, 3) is chosen
     const expectedResult = new Uint8Array([2, 2, 2, 3, 3, 3]);
