+++ app/packages/looker/src/worker/types.ts
@@ -1,0 +1,3 @@
+export interface ReaderMethod {
+  method: string;
+}
